構造体を一つのファイルにまとめた
