+++ backend/go/models/structs.go
@@ -1,5 +1,12 @@
 package models
 
+type Member struct {
+	UserId          string `gorm:"primaryKey"`
+	Name            string `json:"name"`
+	ResistationDate string `json:"resistationDate"`
+	Comment         string `json:"comment"`
+}
+
 type Room struct {
 	RoomId       string   `json:"roomId"`
 	RoomName     string   `json:"roomName"`
